revision provider for deploy resource learns to reconstruct deploy history from F.S.
If the cache file gets wiped out that contains this information, chef
would try to redeploy a revision that has already been deployed, causing
things to fail. This need not be, since we can regenerate this list by
looking at the deployed folders on the filesystem and order them by
ctime.

resolves CHEF-1107